requirements.txt
torch
datasets
numpy
transformers
scikit-learn
huggingface_hub
accelerate
matplotlib
tiktoken
tensorflow
llama-cookbook
ipywidgets
kaggle
torchvision
diffusers[torch]
Pillow
tqdm
datasets
hf_xet


# Backend
fastapi
uvicorn[standard] # Includes standard dependencies like httptools, websockets
pydantic # Often needed implicitly by FastAPI

# Frontend
streamlit
requests # For frontend to call backend


# Optional (if needed by specific models or features)
bitsandbytes # Only if you fix the install for GPU
scipy
ftfy
# Other dependencies...